+++ packages/sit-onyx/src/components/OnyxListbox/OnyxListbox.vue
@@ -4,8 +4,7 @@ import { createComboBox, createId } from "@sit-onyx/headless";
 import { useCheckAll } from "../../composables/checkAll";
 import { computed, ref, watch, watchEffect } from "vue";
 import { useScrollEnd } from "../../composables/scrollEnd";
-import { injectI18n, type OnyxTranslations } from "../../i18n";
-import type { FlattenedKeysOf } from "../../types";
+import { injectI18n } from "../../i18n";
 import OnyxEmpty from "../OnyxEmpty/OnyxEmpty.vue";
 import OnyxListboxOption from "../OnyxListboxOption/OnyxListboxOption.vue";
 import OnyxLoadingIndicator from "../OnyxLoadingIndicator/OnyxLoadingIndicator.vue";
@@ -178,12 +177,15 @@ const { vScrollEnd, isScrollEnd } = useScrollEnd({
   offset: computed(() => props.lazyLoading?.scrollOffset),
 });
 
-const isEmptyMessage = computed<FlattenedKeysOf<OnyxTranslations> | undefined>(
-  () =>
-    (props.options.length === 0 && "listbox.empty") ||
-    (props.withSearch && props.searchTerm && props.options.length === 0 && "listbox.noMatch") ||
-    undefined,
-);
+const isEmptyMessage = computed(() => {
+  if (props.options.length) {
+    return;
+  }
+  if (props.withSearch && props.searchTerm) {
+    return t.value("listbox.noMatch");
+  }
+  return t.value("listbox.empty");
+});
 
 const enabledOptionValues = computed(() =>
   props.options.filter((i) => !i.disabled).map(({ value }) => value),
@@ -246,9 +248,13 @@ watchEffect(() => {
           class="onyx-listbox__search"
         />
 
-        <slot v-if="isEmptyMessage" name="empty" :default-message="t(isEmptyMessage)">
-          <OnyxEmpty>{{ t(isEmptyMessage) }}</OnyxEmpty>
-        </slot>
+        <ul v-if="isEmptyMessage" role="group" aria-label="" class="onyx-listbox__group">
+          <li role="option" aria-selected="false">
+            <slot name="empty" :default-message="isEmptyMessage">
+              <OnyxEmpty>{{ isEmptyMessage }}</OnyxEmpty>
+            </slot>
+          </li>
+        </ul>
 
         <template v-else>
           <ul