+++ packages/sit-onyx/src/index.ts
@@ -74,6 +74,9 @@ export * from "./components/OnyxSwitch/types";
 export { default as OnyxTable } from "./components/OnyxTable/OnyxTable.vue";
 export * from "./components/OnyxTable/types";
 
+export { default as OnyxTextarea } from "./components/OnyxTextarea/OnyxTextarea.vue";
+export * from "./components/OnyxTextarea/types";
+
 export { default as OnyxTooltip } from "./components/OnyxTooltip/OnyxTooltip.vue";
 export * from "./components/OnyxTooltip/types";
 
@@ -89,3 +92,5 @@ export * from "./types";
 
 export type { OnyxTranslations, ProvideI18nOptions } from "./i18n";
 export { createOnyx } from "./utils/plugin";
+
+export { normalizedIncludes } from "./utils/strings";