+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-alpha.60
+
+### Patch Changes
+
+- Updated dependencies [c011e27]
+  - sit-onyx@1.0.0-alpha.105
+
+## 1.0.0-alpha.59
+
+### Patch Changes
+
+- Updated dependencies [20fb878]
+  - sit-onyx@1.0.0-alpha.104
+
+## 1.0.0-alpha.58
+
+### Patch Changes
+
+- Updated dependencies [09ce727]
+  - sit-onyx@1.0.0-alpha.103
+
+## 1.0.0-alpha.57
+
+### Patch Changes
+
+- Updated dependencies [3b9fbf4]
+  - sit-onyx@1.0.0-alpha.102
+
 ## 1.0.0-alpha.56
 
 ### Patch Changes