+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,29 @@
 # sit-onyx
 
+## 1.0.0-alpha.105
+
+### Patch Changes
+
+- c011e27: fix(grid): include padding in max-width
+
+## 1.0.0-alpha.104
+
+### Minor Changes
+
+- 20fb878: feat(theme): include color-scheme in the onyx dark/light theme
+
+## 1.0.0-alpha.103
+
+### Patch Changes
+
+- 09ce727: fix(OnyxCheckbox): align checkbox label padding according to density, stabilize check all border
+
+## 1.0.0-alpha.102
+
+### Minor Changes
+
+- 3b9fbf4: feat: add `OnyxTextarea` component
+
 ## 1.0.0-alpha.101
 
 ### Patch Changes